Fix VR overlay
